add block for create page, and add cart page, update Custom app bar

+++ lib/models/item_model.dart
@@ -12,7 +12,7 @@ class Item with _$Item{
     String? description,
     int? weight,
     int? volume,
-    int? price,
+    double? price,
     String? image,
   }) = _Item;
 }
